# api/api.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
basePath: /
tags:
  - name: deprecated
    description: Old Lagacy API, pleas update ASAP
  - name: privacy-security-evaluator
    description: DITAS privacy security blueprint evaluator
paths:
  /filter:
    post:
      tags:
        - deprecated
        - privacy-security-evaluator
      summary: evaluates a given blueprint request
      operationId: filterPoliciesLagecy
      consumes:
        - application/json
      produces:
        - application/json;charset=UTF-8
        - application/json
        - '*/*'
      parameters:
        - in: body
          name: input
          description: input
          required: true
          schema:
            $ref: '#/definitions/Request'
      responses:
        '200':
          description: ranked blueprint metrics
          schema:
            $ref: '#/definitions/BlueprintRanking'
        '400':
          description: Request input did not match expected format or could not be ranked
  /v1/filter:
    post:
      tags:
        - privacy-security-evaluator
      summary: evaluates a given blueprint request
      operationId: filterPolicies
      consumes:
        - application/json
      produces:
        - application/json;charset=UTF-8
        - application/json
        - '*/*'
      parameters:
        - in: body
          name: input
          description: input
          required: true
          schema:
            $ref: '#/definitions/Request'
      responses:
        '200':
          description: ranked blueprint metrics
          schema:
            $ref: '#/definitions/BlueprintRanking'
        '400':
          description: Request input did not match expected format or could not be ranked
definitions:
  BlueprintRanking:
    type: object
    properties:
      blueprint:
        $ref: '#/definitions/Feature'
      score:
        type: number
        format: double
    example:
      - blueprint:
          id: encryptionAES_128
          type: Encryption
          description: Encryption AES 128
          properties:
            keylength:
              unit: number
              value: 128
            algorithm:
              unit: enum
              value: AES
        score: 1
      - blueprint:
          id: encryptionAES_256
          type: Encryption
          description: Encryption AES 256
          properties:
            keylength:
              unit: number
              value: 256
            algorithm:
              unit: enum
              value: AES
        score: 2

  Feature:
    type: object
    properties:
      description:
        type: string
      id:
        type: string
      properties:
        type: object
        additionalProperties:
          $ref: '#/definitions/Property'
      type:
        type: string
  Property:
    type: object
    properties:
      unit:
        type: string
      value:
        type: object
  Request:
    type: object
    properties:
      blueprintAttributes:
        type: array
        items:
          $ref: '#/definitions/Feature'
      requirement:
        $ref: '#/definitions/Feature'
    example:
      requirement:
        id: encryptionAES_128
        description: Encryption AES 128
        type: Encryption
        properties:
          algorithm:
            unit: enum
            value: AES
          keyLength:
            unit: number
            minimum: 128
      blueprintAttributes:
        - id: encryptionAES_128
          description: Encryption AES 128
          type: Encryption
          properties:
            algorithm:
              unit: enum
              value: AES
            keyLength:
              unit: number
              minimum: 128
        - id: encryptionAES_256
          description: Encryption AES 256
          type: Encryption
          properties:
            algorithm:
              unit: enum
              value: AES
            keyLength:
              unit: number
              minimum: 256